feat: regiseteration mobile ui
