+++ src/components/units/SpotFileInput.tsx
@@ -1,4 +1,10 @@
-import { ForwardedRef, InputHTMLAttributes, forwardRef, useId } from 'react';
+import {
+  ForwardedRef,
+  InputHTMLAttributes,
+  forwardRef,
+  useId,
+  useRef,
+} from 'react';
 import Spot from '@/components/common/Spot';
 import Icon from '@/components/icons';
 
@@ -17,7 +23,7 @@ function SpotFileInput(
     <div key={spotKey} className="flex flex-col items-center gap-[16px]">
       <Spot label={spotKey} />
       <label
-        className="flex h-[121px] w-[175px] items-center justify-center overflow-hidden rounded-[16px] bg-gray-3"
+        className="flex h-[121px] w-[175px] cursor-pointer items-center justify-center overflow-hidden rounded-[16px] bg-gray-3"
         htmlFor={id}
       >
         {previewImgSrc ? (
