+++ src/pages/challenge/components/TagSection.tsx
@@ -1,18 +1,27 @@
 import { twMerge } from 'tailwind-merge';
 import Button from '@/components/common/Button';
+import { useRef } from 'react';
+import { motion } from 'framer-motion';
 
 interface Props {
   className?: string;
 }
 
 function TagSection({ className }: Props) {
+  const contraintsRef = useRef(null);
   return (
-    <div className={twMerge('mx-[21px] flex gap-[8px]', className)}>
-      <Button variant="tag-a">신규</Button>
-      <Button variant="tag-a">인기순</Button>
-      <Button variant="tag-a">추천순</Button>
-      <Button variant="tag-a">도보</Button>
-      <Button variant="tag-a">자전거</Button>
+    <div className="w-full" ref={contraintsRef}>
+      <motion.div
+        drag="x"
+        dragConstraints={contraintsRef}
+        className={twMerge('mx-auto flex w-max gap-[8px]', className)}
+      >
+        <Button variant="tag-a">신규</Button>
+        <Button variant="tag-a">인기순</Button>
+        <Button variant="tag-a">추천순</Button>
+        <Button variant="tag-a">도보</Button>
+        <Button variant="tag-a">자전거</Button>
+      </motion.div>
     </div>
   );
 }
