+++ src/pages/registration/components/MapSection/index.tsx
@@ -1,13 +1,10 @@
 import { useFormContext } from 'react-hook-form';
-import { ChangeEvent } from 'react';
 import KaKaoMap from '@/components/units/KakaoMap';
-import SpotFileInput from '@/components/units/SpotFileInput';
 import { ChallengeRegisterationForm } from '@/types/posts';
-import { getGpsFromImg } from '@/utils/ExifReader';
+import SpotInputList from './SpotInputList';
 
 function MapSection() {
-  const { register, watch, setValue } =
-    useFormContext<ChallengeRegisterationForm>();
+  const { watch } = useFormContext<ChallengeRegisterationForm>();
 
   const spots = watch('spots');
   const spotGps = spots
@@ -31,36 +28,7 @@ function MapSection() {
           사진을 촬영해주세요. (최소 2개, 최대 5개 까지 가능)
         </div>
 
-        <div className="flex justify-between">
-          {[1, 2, 3, 4, 5].map((key, index) => {
-            const spotLabel = `Spot ${key}`;
-            const spot = watch('spots')[index];
-            const file = spot?.files?.item(0);
-            const previewImgSrc = file ? URL.createObjectURL(file) : undefined;
-            const required = index === 0 || index === 1;
-
-            const spotRegister = register(`spots.${index}.files`, {
-              onChange: async (e: ChangeEvent<HTMLInputElement>) => {
-                const { files } = e.target;
-                const targetFile = files?.item(0);
-                if (!targetFile) return;
-                const { lat, lng } = await getGpsFromImg(targetFile);
-                setValue(`spots.${index}.lat`, lat);
-                setValue(`spots.${index}.lng`, lng);
-              },
-              required,
-            });
-
-            return (
-              <SpotFileInput
-                key={key}
-                previewImgSrc={previewImgSrc}
-                spotKey={spotLabel}
-                {...spotRegister}
-              />
-            );
-          })}
-        </div>
+        <SpotInputList />
       </div>
     </>
   );
